requirements.txt
blinker==1.7.0
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.7.0
defusedxml==0.7.1
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
pip-review==1.3.0
pycparser==2.22
python-dotenv==1.0.1
requests==2.31.0
tinycss2==1.3.0
urllib3==2.2.1
visitor==0.1.3
webencodings==0.5.1
Werkzeug==3.0.2
zipp==3.18.1
